requirements.txt
asgiref==3.5.0
async-timeout==4.0.2
autopep8==1.6.0
backports.zoneinfo==0.2.1
Deprecated==1.2.13
Django==4.0.4
django-redis==5.2.0
djangorestframework==3.13.1
packaging==21.3
pycodestyle==2.8.0
PyJWT==2.3.0
pyparsing==3.0.8
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
redis==4.2.2
sqlparse==0.4.2
toml==0.10.2
tzdata==2022.1
tzlocal==4.2
wrapt==1.14.0
